backend/pyproject.toml
[project]
name = "offboardly-backend"
version = "0.1.0"
description = "Offboardly API — AI-powered institutional knowledge capture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.40.0",
    "pinecone-client>=3.0.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "httpx>=0.26.0",
    "markdown>=3.5",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
